requirements.txt
fastapi==0.110.2
passlib==1.7.4
pydantic==2.7.1
pydantic_settings==2.2.1
PyJWT==2.8.0
PyJWT==2.8.0
SQLAlchemy==2.0.29
uvicorn==0.29.0
psycopg2
